# openapi.yaml
openapi: 3.0.0
info:
  title: Game API
  description: API для управления игровой доской
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /game/new:
    get:
      summary: Создает новую доску
      description: Создает и возвращает новую доску, если была старая, то присваивает ей статус досрочного завершения "FINISHED"
      parameters:
        - name: first
          in: query
          description: Определяет, ходит ли игрок первым
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '201':
          description: Новая доска успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDTO'
  /game:
    get:
      summary: Возвращает текущую доску пользователя
      description: Возвращает текущую доску пользователя, если она существует
      responses:
        '200':
          description: Доска успешно возвращена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDTO'
        '404':
          description: Доска не найдена
          content:
            text/plain:
              schema:
                type: string
                example: "Error 404 - Not found: Board not found"
  /game/move:
    get:
      summary: Сделать ход на доске
      description: Сделать ход на доске по указанной позиции
      parameters:
        - name: pos
          in: query
          description: Позиция для хода, значения от 0 до 8
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '201':
          description: Ход успешно сделан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDTO'
        '403':
          description: Некорректный ход
          content:
            text/plain:
              schema:
                type: string
                example: "Error 403 - Forbidden: Illegal move"
        '404':
          description: Доска не найдена
          content:
            text/plain:
              schema:
                type: string
                example: "Error 404 - Not found: Board not found"
  /game/undo:
    get:
      summary: Отменить последний ход
      description: Отменяет последний ход игрока и последний ход машины, или только ход машины если игрок еще не ходил
      responses:
        '200':
          description: Ход успешно отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDTO'
        '403':
          description: Невозможно отменить ход
          content:
            text/plain:
              schema:
                type: string
                example: "Error 403 - Forbidden: Illegal move undo"
        '404':
          description: Доска не найдена
          content:
            text/plain:
              schema:
                type: string
                example: "Error 404 - Not found: Board not found"
components:
  schemas:
    BoardDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор доски
        state:
          type: array
          description: Состояние доски в виде двумерного массива символов
          items:
            type: array
            items:
              type: string
              maxLength: 1
              description: Символ, представляющий состояние клетки (например, 'X', 'O' или пусто)
        status:
          type: string
          description: Статус игры
          enum:
            - IN_PROGRESS
            - FINISHED
            - DRAW
        isUserFirst:
          type: boolean
          description: Определяет, ходит ли игрок первым
        lastMoveTime:
          type: string
          format: date-time
          description: Время последнего хода
      required:
        - id
        - state
        - status
        - isUserFirst
        - lastMoveTime